% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{calc_growth_rates}
\alias{calc_growth_rates}
\title{Calculate growth rates for selected variables}
\usage{
calc_growth_rates(
  dat,
  vars = c("revenue", "epsDiluted", "freeCashFlow"),
  periods = c(10, 5, 3, 1)
)
}
\arguments{
\item{dat}{A object holds financial information, including
balance sheet, income statement, and cash flow.}

\item{vars}{Selected variables}

\item{periods}{Considered periods in years.}
}
\value{
A list with each element corresponding to one variable
}
\description{
This function calculates the growth rates for
the selected variables by \code{vars} in the given
periods by \code{periods}.
}
