% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{calc_health_indice}
\alias{calc_health_indice}
\title{Calculate health index}
\usage{
calc_health_indice(datYear, datQuarter)
}
\arguments{
\item{datYear}{\code{FinData} storing annual values}

\item{datQuarter}{\code{FinData} storing quarterly values, \code{NA}
if no data available.}
}
\value{
A vector with names
}
\description{
Calculate the metrics indicating the health of a company
}
