% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{create_financial_ts}
\alias{create_financial_ts}
\title{Create time series of financial metrics}
\usage{
create_financial_ts(fs_data, metrics, company_cik = NULL)
}
\arguments{
\item{fs_data}{Financial statement data in a standardized format}

\item{metrics}{Character vector of metrics to include in the time series}

\item{company_cik}{CIK of the company to analyze (optional, if NULL analyzes all)}
}
\value{
A list of time series for each requested metric
}
\description{
This function creates time series of financial metrics for trend analysis.
}
