% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{dcf}
\alias{dcf}
\title{Discounted Cash Flow model}
\usage{
dcf(
  profits,
  usePS = FALSE,
  n = 5,
  g = 0.05,
  r = 0.1,
  payout = 0,
  twoStage = F,
  terminalModel = c("multiple", "perpetual"),
  multiple = 10,
  terminalG = 0.05
)
}
\arguments{
\item{profits}{EPS, FCF, etc, the profits generated each year,
if the length is 1, it will be treated as initial value, and
the values for the following years are computed based on growth
rate(s). It can also be sales per share and the paramter \code{usePS}
should be TRUE; in this case, the terminal sales per share will
be computed and stock price will be computed as
terminalSalesPerShare*multiple (assuming multiple is P/S ratio).}

\item{usePS}{logical. If TRUE, the paramter \code{profits} refers to
sales per share and multiple is P/S ratio. Default is FALSE.}

\item{n}{The time span (in years) to consider. This is overriden if
the length of \code{profits} is greater than 1, in which case the number
of years considered is the length of \code{profits}.}

\item{g}{Growth rate. It can be a vector with more than one value,
in which case, each of the rates is considered for each time period
by equally dividing the total time span.}

\item{r}{Discount rate}

\item{payout}{The fraction of profits paid out as dividend.}

\item{twoStage}{Logical, if true, the growth rate will cut to
half for the second half of the considered time span. It is
ignored if \code{g} has more than one value.}

\item{terminalModel}{Model to compute terminal values. 'multiple'
for simply applying a multiple to the value, 'perpetual' for
perpetual growth rate model.}

\item{multiple}{The multiple used for computing terminal values if
the model is \code{multiple}.}

\item{terminalG}{The growth rate used for computing terminal
values if the model is \code{perpetual}.}
}
\description{
Discounted Cash Flow model
}
