% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download_sec_data}
\alias{download_sec_data}
\title{Download SEC financial statement data sets}
\usage{
download_sec_data(
  start_quarter,
  end_quarter = NA,
  outdir = "SEC",
  force = FALSE,
  gap = 5
)
}
\arguments{
\item{start_quarter}{Start quarter in format like "2009q1"}

\item{end_quarter}{End quarter in format like "2009q1". If NA, only downloads start_quarter data.}

\item{outdir}{Directory to save the downloaded files. Default is "SEC".}

\item{force}{Logical. If TRUE, re-download files even if they already exist. Default is FALSE.}
}
\value{
A vector of file paths to the downloaded zip files
}
\description{
This function downloads SEC financial statement data sets from the SEC website.
The data sets are available as quarterly zip files containing sub.txt, num.txt,
tag.txt, and pre.txt files with financial statement data.
}
