% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/edgar_data.R
\docType{data}
\name{edgarData}
\alias{edgarData}
\title{SEC EDGAR data}
\format{
\code{edgarData}
a list containing URLs

\code{edgarData}
a list containing URLs
}
\source{
SEC EDGAR data API

SEC EDGAR data API
}
\usage{
edgarData

edgarData
}
\description{
This object contains the URLs for the source
EDGAR data.

This object contains the URLs for the source
EDGAR data.
}
\keyword{datasets}
