% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{export_to_csv}
\alias{export_to_csv}
\title{Export financial data to CSV}
\usage{
export_to_csv(data, file, ...)
}
\arguments{
\item{data}{The financial data to export}

\item{file}{Path to the output file}

\item{...}{Additional arguments to pass to write.csv}
}
\description{
This function exports financial data to a CSV file.
}
