% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{export_to_excel}
\alias{export_to_excel}
\title{Export financial data to Excel}
\usage{
export_to_excel(data, file, sheet_names = NULL)
}
\arguments{
\item{data}{The financial data to export (can be a single data frame or list of data frames)}

\item{file}{Path to the output Excel file}

\item{sheet_names}{Names for the Excel sheets (if data is a list)}
}
\description{
This function exports financial data to an Excel file.
}
