% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renders.R
\name{find_undervalued_stocks}
\alias{find_undervalued_stocks}
\title{Find undervalued stocks}
\usage{
find_undervalued_stocks(
  metricFile = NULL,
  metrics = NULL,
  sep = ",",
  discount_threshold = 0.8,
  min_market_cap = 100,
  additional_criteria = NULL,
  ...
)
}
\arguments{
\item{metricFile}{The file containing stock metrics}

\item{metrics}{A data.frame containing the data of stock metrics,
if this is provided, the parameter \code{metricFile} will be ignored.}
}
\description{
Reads pre-computed stock metrics on intrinsic values
and add latest stock prices to find the stocks undervalued
}
