% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{get_fx}
\alias{get_fx}
\title{Get current exchange ratio for currencies}
\usage{
get_fx(fromCurrency, toCurrency)
}
\arguments{
\item{fromCurrency}{character vector. The currency to convert from, can accept multiple values}

\item{toCurrency}{character vector. The currency to convert to, can accept multiple values.}
}
\value{
A vector with
}
\description{
This function returns the exchange ratio between \code{fromCurrency} and
\code{toCurrency}. For example, if fromCurrency="HKD" and toCurrency="USD",
a returned value would be kind of 0.128
}
\details{
When one of \code{fromCurrency} and \code{toCurrency} has multiple values, but the other
has only one, then the length-one value will be replicated to match the length
of the longer one.
}
