% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_stock_value}
\alias{get_stock_value}
\title{Calculate the intrinsic value of a stock}
\usage{
get_stock_value(ticker, src = "sa", srcDir = NULL, span = 10, returnDat = F)
}
\arguments{
\item{ticker}{A stock symbol, such as 'AAPL'}

\item{span}{The number of years for consideration.}

\item{returnDat}{Logical. If true, the read data is returned
as attribute 'dat'.}
}
\description{
Calculate the intrinsic value of a stock
}
