% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{get_term_dict}
\alias{get_term_dict}
\title{Get the dictionary between raw and short/standard financial terms}
\usage{
get_term_dict()
}
\description{
Get the dictionary between raw and short/standard financial terms
}
