% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{merge.FinData}
\alias{merge.FinData}
\title{Merge FinData}
\usage{
\method{merge}{FinData}(x, y, mergeType = c("union", "inter"), recentFirst = T, ...)
}
\arguments{
\item{recentFirst}{logical, if TRUE (default), the most recent data will
be listed first.}

\item{a, b}{Two FinData objects.}
}
\description{
Merge FinData
}
