% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_financial_ts}
\alias{plot_financial_ts}
\title{Plot financial time series}
\usage{
plot_financial_ts(
  fs_data,
  metrics,
  company_cik = NULL,
  title = "Financial Metrics Time Series"
)
}
\arguments{
\item{fs_data}{Financial statement data}

\item{metrics}{Character vector of metrics to plot}

\item{company_cik}{CIK of the company to plot (optional, if NULL plots all)}

\item{title}{Plot title}
}
\value{
A ggplot object
}
\description{
This function creates time series plots of financial metrics.
}
