% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ratios}
\alias{plot_ratios}
\title{Plot financial ratios}
\usage{
plot_ratios(
  ratios_data,
  ratio_name,
  company_cik = NULL,
  title = "Financial Ratio Trend"
)
}
\arguments{
\item{ratios_data}{Data frame containing calculated ratios}

\item{ratio_name}{Name of the ratio to plot}

\item{company_cik}{CIK to filter (optional)}

\item{title}{Plot title}
}
\value{
A ggplot object
}
\description{
This function creates visualizations of financial ratios.
}
