% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renders.R
\name{portfolio_value}
\alias{portfolio_value}
\title{Portfolio value}
\usage{
portfolio_value(
  f = NULL,
  dat = NULL,
  tickerCol = "Symbol",
  baseCurrency = "usd",
  ...
)
}
\description{
This function reads a portfolio and return a
a data.frame with current value based on current stock
prices.
}
