% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_balance_sheet}
\alias{read_balance_sheet}
\title{Read balance sheet statement}
\usage{
read_balance_sheet(f, src = c("sa", "mt"))
}
\arguments{
\item{src}{the source of the data, determining how
the file is parsed.}
}
\value{
An object of 'BsData'
}
\description{
Read balance sheet statement
}
