% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_bank_portfolio}
\alias{read_bank_portfolio}
\title{Read bank portfolio}
\usage{
read_bank_portfolio(
  f,
  accountName = NULL,
  bank = c("chase", "cs", "ib"),
  commentChar = "#"
)
}
\description{
Read the portofolio files downloaded from each bank
and return a data.frame with standard columns.
}
