% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_cash_flow}
\alias{read_cash_flow}
\title{Read cash flow statement}
\usage{
read_cash_flow(f, src = c("sa", "mt"))
}
\arguments{
\item{src}{the source of the data, determining how
the file is parsed.}
}
\value{
An object of 'cfData'
}
\description{
Read cash flow statement
}
