% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_financials}
\alias{read_financials}
\title{Read fundamental financial data}
\usage{
read_financials(
  ticker,
  src = c("sa", "mt"),
  period = c("annual", "quarterly"),
  srcDir = NULL
)
}
\arguments{
\item{ticker}{A stock symbol, such as 'AAPL'}
}
\value{
an object of FinData
}
\description{
This function reads the financial data of a company
given a stock symbol
}
