% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_income_statement}
\alias{read_income_statement}
\title{Read income statement}
\usage{
read_income_statement(f, src = c("sa", "mt"))
}
\arguments{
\item{src}{the source of the data, determining how
the file is parsed.}
}
\value{
An object of 'IsData'
}
\description{
Read income statement
}
