% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_sec_num_chunked}
\alias{read_sec_num_chunked}
\title{Read SEC numbers data in chunks (helper function)}
\usage{
read_sec_num_chunked(file, n_max, chunk_size, cols_to_read)
}
\description{
This function reads large files in chunks to manage memory usage.
}
