% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_sec_pre}
\alias{read_sec_pre}
\title{Read SEC presentation data (PRE)}
\usage{
read_sec_pre(file, n_max = -1)
}
\arguments{
\item{file}{Path to the pre.txt file from SEC data set}

\item{n_max}{Number of rows to read (default: all)}
}
\value{
A data frame containing presentation data
}
\description{
This function reads the presentation data from SEC financial statement data sets.
The PRE dataset provides information about how the tags and numbers were presented in financial statements.
}
