% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_sec_sub}
\alias{read_sec_sub}
\title{Read SEC submission data (SUB)}
\usage{
read_sec_sub(file, n_max = -1)
}
\arguments{
\item{file}{Path to the sub.txt file from SEC data set}

\item{n_max}{Number of rows to read (default: all)}
}
\value{
A data frame containing submission data
}
\description{
This function reads the submission data from SEC financial statement data sets.
The SUB dataset contains summary information about an entire EDGAR submission.
}
