% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_sec_tag}
\alias{read_sec_tag}
\title{Read SEC tag data (TAG)}
\usage{
read_sec_tag(file, n_max = -1)
}
\arguments{
\item{file}{Path to the tag.txt file from SEC data set}

\item{n_max}{Number of rows to read (default: all)}
}
\value{
A data frame containing tag data
}
\description{
This function reads the tag data from SEC financial statement data sets.
The TAG dataset includes defining information about each numerical tag.
}
