% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renders.R
\name{reduce_digits}
\alias{reduce_digits}
\title{Reduce digits}
\usage{
reduce_digits(dat, digits = 4)
}
\description{
Reduce the number of digits in the numeric columns
by keeping the specified number of significant digits
}
