% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{standardize_var_names}
\alias{standardize_var_names}
\title{Standardize the variable names}
\usage{
standardize_var_names(varNames)
}
\arguments{
\item{varNames}{A vector of variable names}
}
\value{
updated variable names
}
\description{
Standardize the variable names
}
\keyword{internal}
