% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{stock_screen}
\alias{stock_screen}
\title{Screen stocks based on financial criteria}
\usage{
stock_screen(fs_data, criteria, weights = NULL)
}
\arguments{
\item{fs_data}{Financial statement data}

\item{criteria}{A list of criteria to screen by (e.g., list(roa = list(min = 0.1), pe = list(max = 15)))}

\item{weights}{Optional weights for different criteria in ranking}
}
\value{
A data frame with screened and ranked stocks
}
\description{
This function screens stocks based on specified financial criteria.
}
