% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{summarize_dat}
\alias{summarize_dat}
\title{Summarize data according specified method}
\usage{
summarize_dat(dat, varNames, period, method = "asis", ...)
}
\arguments{
\item{...}{Other arguments to the \code{method}}
}
\value{
A matrix with columns corresponding \code{varNames}.
}
\description{
Summarize data according specified method
}
