% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{summarize_metrics}
\alias{summarize_metrics}
\title{Summarize metrics}
\usage{
summarize_metrics(dat, metrics, removeNegative = T, simplify = F, ...)
}
\description{
Summarizing a metric from the given dataset.
}
