% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tickers}
\alias{tickers}
\title{Stock tickers}
\format{
\subsection{\code{tickers}}{

A data frame with 4 columns:
\itemize{
\item cik: the unique central index key for the ticker.
\item name: company name
\item ticker: ticker symbol
\item exchange: listing stock exchange
}
}
}
\source{
: see \link{edgarData$companyTickersExchange}
}
\usage{
tickers
}
\description{
This dataset contains the stock tickers listed in
US stock exchanges and their company names.
}
\keyword{datasets}
