% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{update_edgar_data}
\alias{update_edgar_data}
\title{Update SEC EDGAR data}
\usage{
update_edgar_data(outDir = NULL)
}
\description{
This function updates the latest versions of the
following files from SEC database.
\itemize{
\item company_tickers_exchange.json
\item company_tickers.json
\item submissions.zip
\item companyfacts.zip
The URLs for these files are in the object \link{edgarData}
}
}
