% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{zz_process_chase_portfolio}
\alias{zz_process_chase_portfolio}
\title{Process portfolio from JP Morgan Chase}
\usage{
zz_process_chase_portfolio(dat)
}
\description{
Process portfolio from JP Morgan Chase
}
\keyword{internal}
