% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{zz_process_ib_portfolio}
\alias{zz_process_ib_portfolio}
\title{Process portfolio from Interactive Broker}
\usage{
zz_process_ib_portfolio(dat)
}
\description{
Process portfolio from Interactive Broker
}
\keyword{internal}
