% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{get_stocks_values}
\alias{get_stocks_values}
\title{Get intrinsic values for multiple stocks}
\usage{
get_stocks_values(tickers, ...)
}
\description{
This is a wrapper of \code{\link[=get_stock_value]{get_stock_value()}} so that
}
