% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_financials}
\alias{read_financials}
\title{Read fundamental financial data}
\usage{
read_financials(
  ticker,
  src = c("sa", "mt", "fmp", "sec"),
  period = c("annual", "quarterly"),
  srcDir = NULL,
  ...
)
}
\arguments{
\item{ticker}{A stock symbol, such as 'AAPL'}

\item{src}{Data source, currently support "sa" (stockanalysis.com) and 'mt' (macrotrends.net)}

\item{srcDir}{As an alternative to src, one can directly provide the folder containing the data files}
}
\value{
an object of FinData
}
\description{
This function reads the financial data of a company
given a stock symbol
}
